Retrive captions from ttml given time stamp+range
This is an example file to get captions from a certain time range
in a ttml file. The file MUST BE PREPROCESSED FIRST to remove any
of the <br> tags and unicode (&39;, etc.).

Then the ttml file is parsed with an xml parser to look for the
div and children <p> elements, each of which is a sentence.i.e.
div[0] is the first sentence spoken in the video.

Given a time of 50s and a range of 10s it performs a binary search
on the ttml file O(logn) looking for the start and end range of the
paragraph elements from time 40s to 60s. Currently it performs this
binary search twice. Once for the starting index (40) and once for the
ending index (60). After the two indexes are prodcued