Binary search for getting captions and times
Need to check for edge cases in the binary search. Also there are problems
with asynchony and getting NoneTypes if the client sends back no data for certain
messages. This causes CRASHES ON CLIENT AND SERVER NEED TO FIX.
Might be a client fix. Otherwise, takes in a time stamp
and returns captions +/- 20s from the stamp.